requirements.txt
secure-smtplib
PyWavelets==1.3.0
numpy
streamlit
pillow==9.4.0
streamlit-image-select==0.5.1
pandas
opencv-python